service_flask/requirements.txt
# Service Flask Core
flask
pyjwt
python-dotenv